split interface for config
