+++ src/config.interface.ts
@@ -1,0 +1,9 @@
+import {AxiosRequestConfig} from 'axios'
+
+export interface IConfig {
+  axios?: AxiosRequestConfig
+  SiteID?: string
+  SitePass?: string
+  ShopID?: string
+  ShopPass?: string
+}
